# backend/travel-go-backend/services/place/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/child-review/list/{review_id}": {
            "get": {
                "description": "Api lấy danh sách đánh giá con",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "Api lấy danh sách đánh giá con",
                "parameters": [
                    {
                        "type": "string",
                        "description": "review_id",
                        "name": "review_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit record for getting",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Current page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
        "/cms/place/list": {
            "get": {
                "description": "Api lấy địa điểm cho cms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cms"
                ],
                "summary": "Api lấy địa điểm cho cms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
...
        "/place/create": {
            "post": {
                "description": "Api tạo địa điểm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "place"
                ],
                "summary": "Api tạo địa điểm",
                "parameters": [
                    {
                        "description": "request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreatePlaceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
        "/place/detail/{id}": {
            "get": {
                "description": "Api lấy chi tiết địa điểm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "place"
                ],
                "summary": "Api lấy chi tiết địa điểm",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
        "/place/favorite": {
            "post": {
                "description": "Api thêm yêu thích địa điểm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorite"
                ],
                "summary": "Api thêm yêu thích địa điểm",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FavoritePlaceAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
        "/place/favorite/list": {
            "get": {
                "description": "Api lấy địa điểm yêu thích",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorite"
                ],
                "summary": "Api lấy địa điểm yêu thích",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit record for getting",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Current page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
        "/place/list": {
            "get": {
                "description": "Api lấy địa điểm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "place"
                ],
                "summary": "Api lấy địa điểm",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
        "/place/update/{id}": {
            "patch": {
                "description": "Api cập nhật địa điểm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "place"
                ],
                "summary": "Api cập nhật địa điểm",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdatePlaceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
        "/review/child/create/{parent_id}": {
            "post": {
                "description": "Api tạo phản hồi đánh giá",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "Api tạo phản hồi đánh giá",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "parent_id",
                        "name": "parent_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ChildReview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
        "/review/create": {
            "post": {
                "description": "Api tạo đánh giá địa điểm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "Api tạo đánh giá địa điểm",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Review"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
        "/review/list/{place_id}": {
            "get": {
                "description": "Api lấy đánh giá địa điểm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "Api lấy đánh giá địa điểm",
                "parameters": [
                    {
                        "type": "string",
                        "description": "place_id",
                        "name": "place_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit record for getting",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Current page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error_info": {
                    "$ref": "#/definitions/api_response.ErrorInfo"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "entity.ChildReview": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "review_img": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ChildReviewImg"
                    }
                }
            }
        },
...
        "request.CreatePlaceReq": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "img": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.CreatePlaceImgReq"
                    }
                },
                "img_checkin": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.CreatePlaceImgCheckinReq"
                    }
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "place_type_id": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        },
...
        "request.FavoritePlaceAddReq": {
            "type": "object",
            "properties": {
                "is_favorite": {
                    "type": "boolean"
                },
                "place_id": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdatePlaceReq": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "img": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.UpdatePlaceImgReq"
                    }
                },
                "img_checkin": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.UpdatePlaceImgCheckinReq"
                    }
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "place_type_id": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        }
    }
}